-changed guid generation for containers and items. Guids will be saved as 32 bit guids instead of 64 bit ones so they dont get screwed. Hopely this fixes any problems regarding missing items. -removed hacky max highguid for items and containers.
WARNING: DO NOT USE THIS REV WITHOUT DATABASE BACKUP!
You need to re-order all item guids to start from 1 to N or you might loose or even fail to create items.

+++ src/game/ItemInterface.cpp
@@ -2289,13 +2289,13 @@ void ItemInterface::mLoadItemsFromDatabase()
 			{
 				if(proto->InventoryType == INVTYPE_BAG)
 				{
-					item=new Container(HIGHGUID_CONTAINER,fields[1].GetUInt64());
+					item=new Container(HIGHGUID_CONTAINER,fields[1].GetUInt32());
 					((Container*)item)->LoadFromDB(fields);
 
 				}
 				else
 				{
-					item = new Item(HIGHGUID_ITEM,fields[1].GetUInt64());
+					item = new Item(HIGHGUID_ITEM,fields[1].GetUInt32());
 					item->LoadFromDB(fields, m_pOwner, false);
 
 				}
